requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements/production.in
#
amqp==5.0.7
    # via kombu
arabic-reshaper==2.1.3
    # via xhtml2pdf
billiard==3.6.4.0
    # via celery
boltons==19.2.0
    # via -r requirements/production.in
boto3==1.12.8
    # via -r requirements/production.in
botocore==1.15.8
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.2.1
    # via
    #   -r requirements/production.in
    #   django-celery-beat
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
defusedxml==0.6.0
    # via odfpy
django==2.2.4
    # via
    #   -r requirements/production.in
    #   django-bootstrap-datepicker-plus
    #   django-cacheops
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-model-utils
    #   django-picklefield
    #   django-redis
    #   django-rest-auth
    #   django-storages
    #   django-timezone-field
django-bootstrap-datepicker-plus==3.0.5
    # via -r requirements/production.in
django-cacheops==4.2
    # via -r requirements/production.in
django-celery-beat==2.2.1
    # via -r requirements/production.in
django-constance[database]==2.4.0
    # via -r requirements/production.in
django-cors-headers==3.1.1
    # via -r requirements/production.in
django-crispy-forms==1.8.0
    # via -r requirements/production.in
django-filter==2.2.0
    # via -r requirements/production.in
django-model-utils==3.2.0
    # via -r requirements/production.in
django-object-actions==1.1.0
    # via -r requirements/production.in
django-picklefield==2.0
    # via
    #   -r requirements/production.in
    #   django-constance
django-redis==4.10.0
    # via -r requirements/production.in
django-rest-auth==0.9.5
    # via -r requirements/production.in
django-storages==1.9.1
    # via -r requirements/production.in
django-timezone-field==4.2.1
    # via django-celery-beat
djangorestframework==3.10.3
    # via
    #   -r requirements/production.in
    #   django-rest-auth
docutils==0.15.2
    # via botocore
drf-writable-nested==0.5.3
    # via -r requirements/production.in
et-xmlfile==1.0.1
    # via openpyxl
factory-boy==2.12.0
    # via -r requirements/production.in
faker==2.0.3
    # via factory-boy
freezegun==0.3.12
    # via -r requirements/production.in
funcy==1.13
    # via django-cacheops
future==0.18.2
    # via arabic-reshaper
html5lib==1.0.1
    # via xhtml2pdf
inflection==0.3.1
    # via -r requirements/production.in
ipy==1.0
    # via -r requirements/production.in
jdcal==1.4.1
    # via openpyxl
jmespath==0.9.5
    # via
    #   boto3
    #   botocore
kombu==5.2.2
    # via celery
markuppy==1.14
    # via tablib
mock==3.0.5
    # via -r requirements/production.in
odfpy==1.4.0
    # via tablib
openpyxl==3.0.0
    # via tablib
pillow==6.2.1
    # via
    #   reportlab
    #   xhtml2pdf
prompt-toolkit==3.0.24
    # via click-repl
psycopg2-binary==2.8.4
    # via -r requirements/production.in
pypdf2==1.26.0
    # via xhtml2pdf
python-bidi==0.4.2
    # via xhtml2pdf
python-crontab==2.3.9
    # via django-celery-beat
python-dateutil==2.8.0
    # via
    #   botocore
    #   faker
    #   freezegun
    #   python-crontab
pytz==2019.3
    # via
    #   -r requirements/production.in
    #   celery
    #   django
    #   django-timezone-field
pyuwsgi==2.0.19.1.post0
    # via -r requirements/production.in
pyyaml==5.1.2
    # via tablib
redis==3.5.3
    # via
    #   -r requirements/production.in
    #   celery
    #   django-cacheops
    #   django-redis
reportlab==3.5.32
    # via xhtml2pdf
s3transfer==0.3.3
    # via boto3
six==1.12.0
    # via
    #   click-repl
    #   django-cacheops
    #   django-rest-auth
    #   faker
    #   freezegun
    #   html5lib
    #   mock
    #   python-bidi
    #   python-dateutil
    #   xhtml2pdf
sqlparse==0.3.0
    # via django
tablib==0.14.0
    # via -r requirements/production.in
text-unidecode==1.3
    # via faker
urllib3==1.25.8
    # via botocore
uwsgitop==0.11
    # via -r requirements/production.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
xhtml2pdf==0.2.5
    # via -r requirements/production.in
xlrd==1.2.0
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
